requirements.txt
autopep8==2.0.2
black==23.3.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.11.0
google-analytics-admin==0.22.5
google-analytics-data==0.18.5
google-api-core==2.17.1
google-auth==2.28.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pipenv==2023.3.20
platformdirs==3.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
requests==2.31.0
rsa==4.9
urllib3==2.2.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
